+++ aws/service.go
@@ -22,6 +22,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/lambda"
 	"github.com/aws/aws-sdk-go/service/organizations"
 	"github.com/aws/aws-sdk-go/service/rds"
+	"github.com/aws/aws-sdk-go/service/route53"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/s3control"
 	"github.com/aws/aws-sdk-go/service/sns"
@@ -343,6 +344,26 @@ func RDSService(ctx context.Context, d *plugin.QueryData, region string) (*rds.R
 	svc := rds.New(sess)
 	d.ConnectionManager.Cache.Set(serviceCacheKey, svc)
 
+	return svc, nil
+}
+
+// Route53Service returns the service connection for AWS route53 service
+func Route53Service(ctx context.Context, connectionManager *connection.Manager, region string) (*route53.Route53, error) {
+	if region == "" {
+		return nil, fmt.Errorf("region must be passed Route53Service")
+	}
+	// have we already created and cached the service?
+	serviceCacheKey := fmt.Sprintf("route53-%s", region)
+	if cachedData, ok := connectionManager.Cache.Get(serviceCacheKey); ok {
+		return cachedData.(*route53.Route53), nil
+	}
+	// so it was not in cache - create service
+	sess, err := getSession(ctx, connectionManager, region)
+	if err != nil {
+		return nil, err
+	}
+	svc := route53.New(sess)
+	connectionManager.Cache.Set(serviceCacheKey, svc)
 	return svc, nil
 }