+++ aws/table_aws_iam_group.go
@@ -67,6 +67,13 @@ func tableAwsIamGroup(_ context.Context) *plugin.Table {
 				Transform:   transform.FromValue(),
 			},
 			{
+				Name:        "inline_policies_std",
+				Description: "Inline policies in canonical form for the group",
+				Type:        proto.ColumnType_JSON,
+				Hydrate:     getAwsIamGroupInlinePolicies,
+				Transform:   transform.FromValue().Transform(inlinePoliciesToStd),
+			},
+			{
 				Name:        "attached_policy_arns",
 				Description: "A list of managed policies attached to the group",
 				Type:        proto.ColumnType_JSON,